requirements.txt
langchain
langchain-community
langchain-openai
python-dotenvstreamlit
